+++ bugsnag-android-core/src/test/java/com/bugsnag/android/InternalEventPayloadDelegateTest.kt
@@ -51,7 +51,8 @@ internal class InternalEventPayloadDelegateTest {
             Notifier()
         )
 
-        val handledState = HandledState.newInstance(HandledState.REASON_HANDLED_EXCEPTION)
+        val handledState = SeverityReason.newInstance(
+            SeverityReason.REASON_HANDLED_EXCEPTION)
         val event = Event(RuntimeException(), config, handledState, NoopLogger)
         delegate.reportInternalBugsnagError(event)